requirements.txt
matplotlib
webcolors
scipy
numpy
Pillow
streamlit
opencv-python
scikit-learn
sklearn
